GROUP-41 merged two local workspaces for frontend to begin work on rendering data for the end user
