--- src/components/Dashboard.css
@@ -1,24 +1,0 @@
-#searchReg{
-    position: absolute;
-    left: 45%;
-    top: 25%;
-    align-items: center;
-    justify-content: center;
-    color:rgb(85, 255, 249) ;
-    font-size: 7mm;
-    font-family: Arial, Helvetica, sans-serifS;
-}
-
-#logOff{
-    position: absolute;
-    right: 2%;
-    top: 5%;
-    align-items: center;
-    justify-content: center;
-}
-
-#eg {
-font-size: 3mm;
-position: absolute;
-left: 28%;
-}
