+++ src/components/Dashboard/Dashboard.css
@@ -11,7 +11,7 @@
 
 #logOff{
     position: absolute;
-    left: 2%;
+    right: 2%;
     top: 5%;
     align-items: center;
     justify-content: center;
