+++ src/components/Home/Home.css
@@ -12,6 +12,36 @@
 
 #logPage{
     position: absolute;
-    left: 28%;
+    left: 23%;
+
+}
+#teamname{
+    
+    color:rgb(85, 255, 249) ;
+    font-size: 7mm;
+    font-family: Arial, Helvetica, sans-serif
+
+}
+#team{
+    position: fixed;
+    left: 45%;
+    top: 5px;
+    justify-content: center;
+    color:rgb(85, 255, 249) ;
+    font-size: 5mm;
+
+}
+#names{
+    position: fixed;
+    word-spacing: 15mm;
+    left: 19.5%;
+    
+    top: 5%;
+    color:rgb(85, 255, 249) ;
+    font-size: 5mm;
+}
+#nav{
+   display: inline;
+    font-size: 6mm;
 
 }
