--- src/components/Map.jsx
@@ -1,26 +1,0 @@
-import { MapContainer, TileLayer, Marker, Popup } from 'react-leaflet'
-import './map.css'
-
-const Map= ()=> {
-
-const position = [51.505, -0.09]
-
-return(
-  <>
-  <MapContainer id= "mapbox" center={position} zoom={13} scrollWheelZoom={false}>
-    <TileLayer
-      attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
-      url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-      
-    />
-    <Marker position={position}>
-      <Popup>
-        A pretty CSS3 popup. <br /> Easily customizable.
-      </Popup>
-    </Marker>
-  </MapContainer>
-  </>
-)
-}
-
-export default Map;
