+++ src/components/NavBar/Navbar.jsx
@@ -1,23 +1,30 @@
 import {Navbar, Nav} from 'react-bootstrap'
+import './home.css'
 
-const SiteNavbar = () => {
+
+
+const Navigation = () => {
 
     return (
-        <>
-        <Navbar className="navbar" bg="dark" variant="dark">
+        <div id = "nav">
+        <Navbar className="navbar" fixed="bottom">
 
-          <Navbar.Brand href="">A bar with whcih to navigate</Navbar.Brand>
-          <Navbar.Toggle aria-controls="basic-navbar-nav" />
-          <Navbar.Collapse id="basic-navbar-nav">
+          <Navbar.Brand href="">Navbar</Navbar.Brand>
+          <Navbar.Toggle aria-controls="navbar-nav" />
+          <Navbar.Collapse id="navbar-nav">
             <Nav className="mr-auto">
-              <Nav.Link href="/">Home</Nav.Link>
-              <Nav.Link href="/Dashboard">Search Some Crims</Nav.Link>
+              <Nav.Link href="/">Home</Nav.Link> 
+              <br/>
+              <br/>
+              <Nav.Link href="/dashboard">Dashboard</Nav.Link>
             </Nav>
 
           </Navbar.Collapse>
         </Navbar>      
-        </>
+        </div>
+    
+
     )
 }
 
-export default SiteNavbar
+export default Navigation;
