--- src/components/Results.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import Map from './Map';
-
-    
-
-function Results() {
-    
-    return (
-      <div>
-       <Map/>
-      </div>
-    );
-  }
-    export default Results;
