--- src/components/map.css
@@ -1,12 +1,0 @@
-#mapbox{
-    position: fixed;
-    height: 125mm;
-    width: 100%;
-    border: white;
-    right: 1mm;
-    bottom: 10px;
-    
-    
- }
- 
-
